# Conflicts:
#	src/main/java/com/spring/jwt/Interfaces/IInvoice1.java
#	src/main/java/com/spring/jwt/controller/Invoice1Controller.java
#	src/main/java/com/spring/jwt/controller/ProductsController.java
#	src/main/java/com/spring/jwt/entity/Products.java
#	src/main/java/com/spring/jwt/service/Invoice1Service.java
#	src/main/resources/application-dev.properties

+++ src/main/java/com/spring/jwt/dto/ProductsDTO.java
@@ -20,7 +20,7 @@ public class ProductsDTO {
 
     private ClothingType clothingType;
 
-    private List<Integer>stockQuantities;
+    private List<Double>stockQuantities;
 
 }
